restructured samplers into seperate module & class files
