--- src/eva/vision/data/wsi/patching/samplers.py
@@ -1,236 +1,0 @@
-"""Samplers for WSI patch extraction."""
-
-import abc
-import random
-from typing import Generator, Tuple
-
-import numpy as np
-
-from eva.vision.data.wsi.patching.mask import Mask
-
-
-class Sampler(abc.ABC):
-    """Base class for samplers."""
-
-    @abc.abstractmethod
-    def sample(
-        self,
-        width: int,
-        height: int,
-        layer_shape: Tuple[int, int],
-        mask: Mask | None = None,
-    ) -> Generator[Tuple[int, int], None, None]:
-        """Sample patche coordinates.
-
-        Args:
-            width: The width of the patches.
-            height: The height of the patches.
-            layer_shape: The shape of the layer.
-            mask: Tuple containing the mask array and the scaling factor with respect to the
-                provided layer_shape. Optional, only required for samplers with foreground
-                filtering.
-
-        Returns:
-            A generator producing sampled patch coordinates.
-        """
-
-
-class ForegroundSampler(Sampler):
-    """Base class for samplers with foreground filtering capabilities."""
-
-    @abc.abstractmethod
-    def is_foreground(
-        self,
-        mask: Mask,
-        x: int,
-        y: int,
-        width: int,
-        height: int,
-        min_foreground_ratio: float,
-    ) -> bool:
-        """Check if a patch contains sufficient foreground."""
-
-
-class RandomSampler(Sampler):
-    """Sample patch coordinates randomly.
-
-    Args:
-        n_samples: The number of samples to return.
-        seed: The random seed.
-    """
-
-    def __init__(self, n_samples: int = 1, seed: int = 42):
-        """Initializes the sampler."""
-        self.seed = seed
-        self.n_samples = n_samples
-
-    def sample(
-        self,
-        width: int,
-        height: int,
-        layer_shape: Tuple[int, int],
-    ) -> Generator[Tuple[int, int], None, None]:
-        """Sample random patches.
-
-        Args:
-            width: The width of the patches.
-            height: The height of the patches.
-            layer_shape: The shape of the layer.
-        """
-        _set_seed(self.seed)
-
-        for _ in range(self.n_samples):
-            x_max, y_max = layer_shape[0], layer_shape[1]
-            x, y = random.randint(0, x_max - width), random.randint(0, y_max - height)  # nosec
-            yield x, y
-
-
-class GridSampler(Sampler):
-    """Sample patches based on a grid.
-
-    Args:
-        max_samples: The maximum number of samples to return.
-        overlap: The overlap between patches in the grid.
-        seed: The random seed.
-    """
-
-    def __init__(
-        self,
-        max_samples: int | None = None,
-        overlap: Tuple[int, int] = (0, 0),
-        seed: int = 42,
-    ):
-        """Initializes the sampler."""
-        self.max_samples = max_samples
-        self.overlap = overlap
-        self.seed = seed
-
-    def sample(
-        self,
-        width: int,
-        height: int,
-        layer_shape: Tuple[int, int],
-    ) -> Generator[Tuple[int, int], None, None]:
-        """Sample patches from a grid.
-
-        Args:
-            width: The width of the patches.
-            height: The height of the patches.
-            layer_shape: The shape of the layer.
-        """
-        x_y, indices = _get_grid_coords_and_indices(layer_shape, width, height, self.overlap)
-        max_samples = len(indices) if self.max_samples is None else self.max_samples
-        for i in indices[:max_samples]:
-            yield x_y[i]
-
-
-class ForegroundGridSampler(ForegroundSampler):
-    """Sample patches based on a grid, only returning patches containing foreground.
-
-    Args:
-        max_samples: The maximum number of samples to return.
-        overlap: The overlap between patches in the grid.
-        min_foreground_ratio: The minimum amount of foreground within a sampled patch.
-        seed: The random seed.
-    """
-
-    def __init__(
-        self,
-        max_samples: int = 20,
-        overlap: Tuple[int, int] = (0, 0),
-        min_foreground_ratio: float = 0.35,
-        seed: int = 42,
-    ):
-        """Initializes the sampler."""
-        self.max_samples = max_samples
-        self.overlap = overlap
-        self.min_foreground_ratio = min_foreground_ratio
-        self.seed = seed
-
-    def sample(
-        self,
-        width: int,
-        height: int,
-        layer_shape: Tuple[int, int],
-        mask: Mask,
-    ):
-        """Sample patches from a grid containing foreground.
-
-        Args:
-            width: The width of the patches.
-            height: The height of the patches.
-            layer_shape: The shape of the layer.
-            mask: The mask of the image.
-        """
-        x_y, indices = _get_grid_coords_and_indices(layer_shape, width, height, self.overlap)
-
-        count = 0
-        for i in indices:
-            if count >= self.max_samples:
-                break
-            if self.is_foreground(
-                mask, x_y[i][0], x_y[i][1], width, height, self.min_foreground_ratio
-            ):
-                count += 1
-                yield x_y[i]
-
-    def is_foreground(
-        self,
-        mask: Mask,
-        x: int,
-        y: int,
-        width: int,
-        height: int,
-        min_foreground_ratio: float,
-    ) -> bool:
-        """Check if a patch contains sufficient foreground.
-
-        Args:
-            mask: The mask of the image.
-            x: The x-coordinate of the patch.
-            y: The y-coordinate of the patch.
-            width: The width of the patch.
-            height: The height of the patch.
-            min_foreground_ratio: The minimum amount of foreground in the patch.
-        """
-        x_, y_ = self._scale_coords(x, y, mask.scale_factors)
-        width_, height_ = self._scale_coords(width, height, mask.scale_factors)
-        patch_mask = mask.mask_array[y_ : y_ + height_, x_ : x_ + width_]
-        return patch_mask.sum() / patch_mask.size > min_foreground_ratio
-
-    def _scale_coords(self, x: int, y: int, scale_factors: Tuple[float, float]) -> Tuple[int, int]:
-        return int(x / scale_factors[0]), int(y / scale_factors[1])
-
-
-def _get_grid_coords_and_indices(
-    layer_shape: Tuple[int, int],
-    width: int,
-    height: int,
-    overlap: Tuple[int, int],
-    shuffle: bool = True,
-    seed: int = 42,
-):
-    """Get grid coordinates and indices.
-
-    Args:
-        layer_shape: The shape of the layer.
-        width: The width of the patches.
-        height: The height of the patches.
-        overlap: The overlap between patches in the grid.
-        shuffle: Whether to shuffle the indices.
-        seed: The random seed.
-    """
-    x_range = range(0, layer_shape[0] - width, width - overlap[0])
-    y_range = range(0, layer_shape[1] - height, height - overlap[1])
-    x_y = [(x, y) for x in x_range for y in y_range]
-
-    indices = list(range(len(x_y)))
-    if shuffle:
-        _set_seed(seed)
-        np.random.shuffle(indices)
-    return x_y, indices
-
-
-def _set_seed(seed: int) -> None:
-    random.seed(seed)
-    np.random.seed(seed)
